+++ tests/Jroc.Tests/Node/Dns/GeneratorTests.cs
@@ -1,0 +1,15 @@
+using System.Threading.Tasks;
+
+namespace Jroc.Tests.Node.Dns;
+
+public class GeneratorTests : GeneratorTestsBase
+{
+    public GeneratorTests()
+        : base("Node/Dns")
+    {
+    }
+
+    [Fact]
+    public Task Require_Dns_Lookup()
+        => GenerateTest(nameof(Require_Dns_Lookup));
+}